+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/permission/MiniAppDownloadedListAdapter.kt
@@ -1,0 +1,54 @@
+package com.rakuten.tech.mobile.testapp.ui.permission
+
+import android.net.Uri
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import com.rakuten.tech.mobile.miniapp.MiniAppInfo
+import com.rakuten.tech.mobile.miniapp.testapp.databinding.DownloadedItemListMiniappBinding
+import com.rakuten.tech.mobile.testapp.helper.setIcon
+
+class MiniAppDownloadedListAdapter(private val miniAppListener: MiniAppDownloadedListener) :
+    RecyclerView.Adapter<MiniAppDownloadedListAdapter.ViewHolder?>() {
+    private var miniApps: List<MiniAppInfo> = listOf()
+    private var miniAppPermissions: List<String> = listOf()
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val layoutInflater = LayoutInflater.from(parent.context)
+        val binding = DownloadedItemListMiniappBinding.inflate(layoutInflater, parent, false)
+
+        return ViewHolder(binding)
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        setIcon(holder.miniAppRoot.context, Uri.parse(miniApps[position].icon), holder.miniAppIcon)
+        holder.miniAppName.text = miniApps[position].displayName
+        holder.miniAppPermissions.text = miniAppPermissions[position]
+        holder.miniAppRoot.setOnClickListener {
+            miniAppListener.onMiniAppItemClick(miniApps[position])
+        }
+    }
+
+    override fun getItemCount(): Int = miniApps.size
+
+    fun addDownloadedList(apps: List<MiniAppInfo>, permissions: List<String>) {
+        miniApps = apps
+        miniAppPermissions = permissions
+        notifyDataSetChanged()
+    }
+
+    interface MiniAppDownloadedListener {
+        fun onMiniAppItemClick(miniAppInfo: MiniAppInfo)
+    }
+
+    inner class ViewHolder(itemView: DownloadedItemListMiniappBinding) :
+        RecyclerView.ViewHolder(itemView.root) {
+        val miniAppRoot: View = itemView.downloadedMiniappRoot
+        val miniAppIcon: ImageView = itemView.downloadedMiniappIcon
+        val miniAppName: TextView = itemView.downloadedMiniappName
+        val miniAppPermissions: TextView = itemView.downloadedMiniappPermissions
+    }
+}